requirements.txt
fastapi==0.109.2
uvicorn==0.29.0
pydantic==2.7.0
pandas==2.0.3
numpy==1.26.4
requests==2.31.0
streamlit==1.32.0
scikit-learn==1.6.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.6
python-dotenv==1.0.1